Move loss into its own module
